requirements.txt
flask
pytesseract
opencv-python
pandas
pillow
pdf2image
